move test outside pytestmark file
