+++ tests/integration/test_regression_mol.py
@@ -9,7 +9,6 @@ from torch.utils.data import DataLoader
 
 from chemprop import nn
 from chemprop.data import MoleculeDatapoint, MoleculeDataset, collate_batch
-from chemprop.models import MPNN
 
 pytestmark = [
     pytest.mark.parametrize(
@@ -76,49 +75,3 @@ def test_overfit(mpnn, dataloader):
     mse = errors.square().mean().item()
 
     assert mse <= 0.05
-
-
-def test_output_transform(data):
-    train_dset = MoleculeDataset(data)
-    output_scaler = train_dset.normalize_targets()
-    train_loader = DataLoader(train_dset, 32, collate_fn=collate_batch)
-
-    test_dset = MoleculeDataset(data)
-    test_loader = DataLoader(test_dset, 32, collate_fn=collate_batch, shuffle=False)
-
-    output_transform = nn.UnscaleTransform.from_standard_scaler(output_scaler)
-    ffn = nn.RegressionFFN(output_transform=output_transform)
-    mpnn = MPNN(nn.BondMessagePassing(), nn.MeanAggregation(), ffn)
-
-    trainer = pl.Trainer(
-        logger=False,
-        enable_checkpointing=False,
-        enable_progress_bar=True,
-        enable_model_summary=False,
-        accelerator="cpu",
-        devices=1,
-        max_epochs=100,
-        overfit_batches=1.00,
-    )
-    trainer.fit(mpnn, train_loader)
-
-    mpnn.train()
-    predss = []
-    for batch in train_loader:
-        bmg, _, _, targets, *_ = batch
-        preds = mpnn(bmg)
-        predss.append(preds)
-
-    preds = torch.cat(predss)
-    std, mean = torch.std_mean(preds, dim=0)
-
-    assert torch.allclose(std, torch.ones_like(std), atol=0.1)
-    assert torch.allclose(mean, torch.zeros_like(mean), atol=0.1)
-
-    predss = trainer.predict(mpnn, test_loader)
-    preds = torch.cat(predss)
-    std, mean = torch.std_mean(preds, dim=0)
-    y_std, y_mean = torch.std_mean(torch.from_numpy(test_dset.Y), dim=0)
-
-    assert torch.allclose(std, y_std, atol=0.1)
-    assert torch.allclose(mean, y_mean, atol=0.1)
